pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reasoning-core"
version = "0.1.0"
description = "Universal reasoning extraction engine - Transform expertise into intelligent knowledge graphs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Excelsior2026", email = "billparris@gmail.com"}
]
keywords = [
    "reasoning",
    "knowledge-graphs",
    "nlp",
    "concept-extraction",
    "ai",
    "machine-learning",
    "education",
    "training"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Excelsior2026/reasoning-core"
Repository = "https://github.com/Excelsior2026/reasoning-core"
Issues = "https://github.com/Excelsior2026/reasoning-core/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
